backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pgvector==0.2.4
groq==0.4.2
google-generativeai==0.4.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
